requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3
numpy==1.26.3
Pillow==10.2.0
pytest==8.0.0
httpx==0.26.0
pytest-cov==5.0.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0